Fixes.
Fix for issue #7 .
Fixed the bug where if someone that is not you knocked down a NPC, then that would be not shown in your client.
Fixed the issue with abnormalities where if you have mode 2 or above enabled, you wouldn't be able to see other members debuffs on the target.
Fix for the contracts ids change.